requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.21
langchain_huggingface==0.1.2
langchain-ollama==0.2.0
pypdf==5.0.1
docx2txt==0.8
faiss-cpu==1.9.0
sqlalchemy==2.0.35
python-multipart==0.0.12
pytest==8.3.3
ollama
streamlit
python-dotenv
